+++ priv/repo/migrations/20221130093126_add_subscope_to_template.exs
@@ -1,0 +1,9 @@
+defmodule TdDf.Repo.Migrations.AddSubscopeToTemplate do
+  use Ecto.Migration
+
+  def change do
+    alter table("templates") do
+      add :subscope, :string, null: true
+    end
+  end
+end